backend/requirements.txt
fastapi
uvicorn
psycopg2-binary==2.9.10
pandas